control flow graph refactor
refactor the control flow graph. The biggest change is that it is now
purely concerned with the control flow inside of functions, thus, the
new method CFGraph::parse_program(data: ast::ProgramData), will return a
Vec of CFGraphs, where each Graph is responsible for a single function.
This approach greatly simplifies the CFG implementation details, as the
CFG is now responsible for a much smaller scope of operations (it's
entire responsibility is now to analyse and optimize the control flow of
single functions)